requirements.txt
pycryptodome
six
python-decouple
requests~=2.24.0
django~=3.1.2
django-cors-headers
djangorestframework==3.11.2
pytz
sqlparse
coreapi
django-filter~=2.4.0
django-simpleui
mysqlclient
redis
celery
django-redis
drf-yasg
sentry-sdk
python-dateutil
django-debug-toolbar~=3.1.1
packaging
python-alipay-sdk
django-json-widget